refactor(controlPanel_SearchBar_ContactForm)
